KeepLogsOnSuccess

+++ test.ps1
@@ -30,6 +30,9 @@ This script provides helpers for running executing the MsQuic tests.
 .PARAMETER ConvertLogs
     Will also convert logs to text. Only works when LogProfile is set.
 
+.PARAMETER KeepLogsOnSuccess
+    Don't discard logs on success.
+
 .EXAMPLE
     test.ps1
 
@@ -81,7 +84,10 @@ param (
     [switch]$Debugger = $false,
 
     [Parameter(Mandatory = $false)]
-    [switch]$ConvertLogs = $false
+    [switch]$ConvertLogs = $false,
+
+    [Parameter(Mandatory = $false)]
+    [switch]$KeepLogsOnSuccess = $false
 )
 
 Set-StrictMode -Version 'Latest'
@@ -281,11 +287,22 @@ function FinishTestCase($TestCase) {
     Add-XmlResults $TestCase
 
     if ($Debugger -or $stdout.Contains("[  PASSED  ] 1 test")) {
-        if ($LogProfile -ne "None") {
-            # Don't keep logs on success.
-            .\log.ps1 -Cancel -InstanceName $TestCase.InstanceName | Out-Null
+        if ($KeepLogsOnSuccess) {
+            if ($LogProfile -ne "None") {
+                # Keep logs on failure.
+                if ($ConvertLogs) {
+                    .\log.ps1 -Stop -OutputDirectory $TestCase.LogDir -InstanceName $TestCase.InstanceName -ConvertToText | Out-Null
+                } else {
+                    .\log.ps1 -Stop -OutputDirectory $TestCase.LogDir -InstanceName $TestCase.InstanceName | Out-Null
+                }
+            }
+        } else {
+            if ($LogProfile -ne "None") {
+                # Don't keep logs on success.
+                .\log.ps1 -Cancel -InstanceName $TestCase.InstanceName | Out-Null
+            }
+            Remove-Item $TestCase.LogDir -Recurse -Force | Out-Null
         }
-        Remove-Item $TestCase.LogDir -Recurse -Force | Out-Null
     } else {
         if ($LogProfile -ne "None") {
             # Keep logs on failure.
@@ -397,7 +414,7 @@ try {
     Log "$($TestCount) test(s) run. $($TestsFailed) test(s) failed."
     if ($TestsFailed -ne 0) {
         Log "Logs can be found in $($LogDir)"
-    } else {
+    } elseif (!$KeepLogsOnSuccess) {
         Remove-Item $LogDir -Recurse -Force | Out-Null
     }
 }
